+++ tests/docs/test_token_diffs.py
@@ -1,6 +1,7 @@
 from textwrap import dedent
 
-from flexdoc.docs.text_doc import SentIndex, TextDoc
+from flexdoc.docs import SentIndex
+from flexdoc.docs.text_doc import TextDoc
 from flexdoc.docs.token_diffs import DiffStats, diff_wordtoks, find_best_alignment
 
 _short_text1 = dedent(